NP-46072 Refactor ListPagination component to wrap lists
