+++ src/pages/basic_data/app_admin/central_import/CentralImportPage.tsx
@@ -4,10 +4,10 @@ import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import {
+  fetchImportCandidates,
   FetchImportCandidatesParams,
   ImportCandidateOrderBy,
   SortOrder,
-  fetchImportCandidates,
 } from '../../../../api/searchApi';
 import { ErrorBoundary } from '../../../../components/ErrorBoundary';
 import { ListPagination } from '../../../../components/ListPagination';
@@ -94,7 +94,12 @@ export const CentralImportPage = ({ statusFilter, yearFilter }: CentralImportPag
       {importCandidateQuery.isLoading ? (
         <ListSkeleton minWidth={100} maxWidth={100} height={100} />
       ) : searchResults.length > 0 ? (
-        <>
+        <ListPagination
+          count={importCandidateQuery.data?.size ?? 0}
+          rowsPerPage={rowsPerPage}
+          page={page + 1}
+          onPageChange={(newPage) => updatePath(((newPage - 1) * rowsPerPage).toString(), rowsPerPage.toString())}
+          onRowsPerPageChange={(newRowsPerPage) => updatePath('0', newRowsPerPage.toString())}>
           <List>
             {searchResults.map((importCandidate) => (
               <ErrorBoundary key={importCandidate.id}>
@@ -102,14 +107,7 @@ export const CentralImportPage = ({ statusFilter, yearFilter }: CentralImportPag
               </ErrorBoundary>
             ))}
           </List>
-          <ListPagination
-            count={importCandidateQuery.data?.size ?? 0}
-            rowsPerPage={rowsPerPage}
-            page={page + 1}
-            onPageChange={(newPage) => updatePath(((newPage - 1) * rowsPerPage).toString(), rowsPerPage.toString())}
-            onRowsPerPageChange={(newRowsPerPage) => updatePath('0', newRowsPerPage.toString())}
-          />
-        </>
+        </ListPagination>
       ) : (
         <Typography sx={{ mt: '1rem' }}>{t('common.no_hits')}</Typography>
       )}
