+++ src/pages/editor/EditorCurators.tsx
@@ -51,7 +51,15 @@ export const EditorCurators = () => {
       <i>{t('editor.curators.no_users_found')}</i>
     </Typography>
   ) : (
-    <>
+    <ListPagination
+      count={curators.length}
+      rowsPerPage={rowsPerPage}
+      page={validPage}
+      onPageChange={(newPage) => setPage(newPage)}
+      onRowsPerPageChange={(newRowsPerPage) => {
+        setRowsPerPage(newRowsPerPage);
+        setPage(1);
+      }}>
       <TableContainer component={Paper} sx={{ mb: '0.5rem' }}>
         <Table size="small" sx={alternatingTableRowColor}>
           <TableHead>
@@ -68,16 +76,6 @@ export const EditorCurators = () => {
           </TableBody>
         </Table>
       </TableContainer>
-      <ListPagination
-        count={curators.length}
-        rowsPerPage={rowsPerPage}
-        page={validPage}
-        onPageChange={(newPage) => setPage(newPage)}
-        onRowsPerPageChange={(newRowsPerPage) => {
-          setRowsPerPage(newRowsPerPage);
-          setPage(1);
-        }}
-      />
-    </>
+    </ListPagination>
   );
 };
