+++ src/pages/search/registration_search/RegistrationSearch.tsx
@@ -2,7 +2,6 @@ import { Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { ListPagination } from '../../../components/ListPagination';
-import { ListPaginationTop } from '../../../components/ListPaginationTop';
 import { ListSkeleton } from '../../../components/ListSkeleton';
 import { ROWS_PER_PAGE_OPTIONS } from '../../../utils/constants';
 import { SearchParam } from '../../../utils/searchHelpers';
@@ -32,8 +31,6 @@ export const RegistrationSearch = ({ registrationQuery }: Pick<SearchPageProps, 
         <ListSkeleton arrayLength={3} minWidth={40} height={100} />
       ) : registrationQuery.data?.hits && registrationQuery.data.hits.length > 0 ? (
         <>
-          <ListPaginationTop count={registrationQuery.data.totalHits} page={page + 1} rowsPerPage={rowsPerPage} />
-          <RegistrationSearchResults searchResult={registrationQuery.data.hits} />
           <ListPagination
             count={registrationQuery.data.totalHits}
             page={page + 1}
@@ -41,7 +38,9 @@ export const RegistrationSearch = ({ registrationQuery }: Pick<SearchPageProps, 
             rowsPerPage={rowsPerPage}
             onRowsPerPageChange={(newRowsPerPage) => updatePath('0', newRowsPerPage.toString())}
             maxHits={10_000}
-          />
+            showPaginationTop={true}>
+            <RegistrationSearchResults searchResult={registrationQuery.data.hits} />
+          </ListPagination>
         </>
       ) : (
         <Typography sx={{ mx: { xs: '0.5rem', md: 0 } }}>{t('common.no_hits')}</Typography>
